# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/calculator": {
            "post": {
                "description": "Calculates the optimal pack combination for a requested amount.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Calculate packs by amount",
                "parameters": [
                    {
                        "description": "Calculator payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodel.CalculatorPacksRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.CalculatorPacksResponse"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the current API health status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check API health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.HealthCheckResponse"
                        }
                    }
                }
            }
        },
        "/packs": {
            "get": {
                "description": "Returns all configured package sizes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packs"
                ],
                "summary": "List pack sizes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.GetPacksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.HttpErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Replaces all configured package sizes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packs"
                ],
                "summary": "Update pack sizes",
                "parameters": [
                    {
                        "description": "Pack sizes payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodel.UpdateAllPacksRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.HttpErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "viewmodel.CalculatorPacksRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 500000
                }
            }
        },
        "viewmodel.CalculatorPacksResponse": {
            "type": "object",
            "properties": {
                "pack_size": {
                    "type": "integer",
                    "example": 23
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "viewmodel.GetPacksResponse": {
            "type": "object",
            "properties": {
                "pack_sizes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "viewmodel.HealthCheckResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "viewmodel.HttpErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid-request-body"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.validationError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "some parameter in the request body is invalid"
                },
                "status": {
                    "type": "integer",
                    "example": 422
                }
            }
        },
        "viewmodel.UpdateAllPacksRequest": {
            "type": "object",
            "properties": {
                "pack_sizes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "viewmodel.validationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "must be a valid quantity with value greater than 0"
                },
                "parameter": {
                    "type": "string",
                    "example": "quantity"
                }
            }
        }
    }
}